RF: Move clone-internal helper to clone.py
Not used elsewhere, not actually useful in a generic context. Also
move the associated tests to improve modularity.
